wallet-generator/requirements.txt
web3==6.8.0 
eth-account==0.6.0  
mnemonic==0.21
google-cloud-firestore==2.21.0
fastapi==0.75.0  
uvicorn==0.17.6
